Move `transport/whisper/gethbridge` to `bridge/geth`
