+++ bridge/geth/filter.go
