+++ bridge/geth/message_store.go
