+++ bridge/geth/public_whisper_api.go
