+++ bridge/geth/syncevent_response.go
